add new change

+++ src/task3/Food.java
@@ -2,21 +2,21 @@ package task3;
 
 public enum Food {
 
-   APPLE(5), POTATO(3), MILK(20),BEER(15), TOBACCO(30);
+   APPLE(5.0), POTATO(3.0), MILK(20.0),BEER(15.0), TOBACCO(30.0);
 
-   public int cost;
+   public double cost;
 
 
-   public int getCost() {
+   public double getCost() {
       return cost;
    }
 
-   public void setCost(int cost) {
+   public void setCost(double cost) {
       this.cost = cost;
    }
 
 
-   Food(int cost) {
+   Food(double cost) {
       this.cost = cost;
    }
 }
